requirements.txt
python-dotenv
pandas
pywinauto
decouple